Changed the database to remove a main index, which I had put as a mistake. Note from next time to drop indexes which are not required
By indexes I mean the name of the model key for a variable. If we do a modelVar.save() it will create the model in the database. Any subsequent local change in the model structure won't be reflected in the server. Thus go to mognodb.com to make the model structure changes to allow subsequent saves/updates to the database for that model.